Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 
 int main() {
-    float a,b; 
+    int a,b; 
     char c;
-    scanf("%f", &a);
-    scanf("%f", &b);
+    scanf("%d", &a);
+    scanf("%d", &b);
     scanf("%c", &c);
     if(c='+'){
         printf("%d",a+b);
